requirements.txt
pyftpdlib~=1.5.7
dnslib~=0.9.23
dnspython~=2.4.2